Synchronize deletion of mInterface in NSFW destructor
We're seeing intermittent crashes on Windows calling 
mInterface->isWatching() in StartWorker::Execute on a background thread. 
The only unsynchronized access to mInterface that we can find takes 
place in the destructor of NSFW itself, making us wonder if interleaved 
calls to start and stop might be causing the following race condition.

* Start is called and sets the persistent handle
* Stop is called
* Before the StopWorker's HandleOkCallback runs and clears the 
persistent handle, we call start again
* Start spawns a worker thread
* Stop's ok callback clears the persistent handle
* The garbage collector runs and runs the destructor on the NSFW object, 
which clears the mInterface field
* The thread accesses the mInterface pointer, which has now been deleted

We don't have proof that this is what is happening, but if it is, I'm 
hoping that locking the field access in the destructor would prevent it.